+++ take/__init__.py
@@ -1,4 +1,4 @@
 
-__version__ = '0.0.4'
+__version__ = '0.0.5'
 # main entry point
 from .take_template import TakeTemplate