+++ take/take_template.py
@@ -5,6 +5,11 @@ from .parser import parse
 
 class TakeTemplate(object):
 
+    @staticmethod
+    def from_file(path):
+        with open(path, 'rb') as f:
+            return TakeTemplate(f.read().decode('utf-8'))
+
     def __init__(self, src, **kwargs):
         self.node = parse(src)
         self.base_url = kwargs.get('base_url', None)